Install servicemenus in new install location
Since KIO 5.85, this new file location is preferred over the
KServiceTypeTrader results.

With this change, single desktop files from the KDE store are installed
in the correct location. If however a custom install script is used,
developers need to adjust their scripts.

Because the "kio/servicemenus" locations is different than the one
KAuthorized allows by default, we have to mark the file as executable.

Task: https://phabricator.kde.org/T14543